Cleanup build logic of obs-browser
obs-frontend-api had limited use while being an extra dependency
that was frontend specific. As a result, I've removed it. However,
this still needs to be done for MacOS, the code will not compile on
MacOS.

I've removed the requirement of debug binaries in the libcef
distribution. The most common distribution given by spotifiy has
a minimal distribution which removes tests and debug binaries. Debug
binaries are still detected and given under the variable
CEF_DEBUG_LIBRARIES though it should be tested for right now if
you wish to use it.

Instead of post build commands, we use install commands which is
far more consistent. We also install into the plugins directory
specified by the helper variables. As a result, it now works with
cpack and install commands. This still needs to be done for MacOS.